Use DI for AppConfig & KeyVault in Webhook Router.
